Ported new control flow over to tpcf_jackknife
