+++ halotools/mock_observables/two_point_clustering/clustering_helpers.py
@@ -15,8 +15,7 @@ from ..mock_observables_helpers import enforce_sample_has_correct_shape
 from ...custom_exceptions import HalotoolsError
 from ...utils.array_utils import convert_to_ndarray, array_is_monotonic
 
-__all__ = ('_tpcf_jackknife_process_args',
-    '_s_mu_tpcf_process_args',
+__all__ = ('_s_mu_tpcf_process_args',
     '_marked_tpcf_process_args', '_tpcf_one_two_halo_decomp_process_args')
 
 __author__=['Duncan Campbell', 'Andrew Hearin']
@@ -109,145 +108,6 @@ def downsample_inputs_exceeding_max_sample_size(sample1, sample2, _sample1_is_sa
             pass
 
     return sample1, sample2
-
-
-def _tpcf_jackknife_process_args(sample1, randoms, rbins, 
-    Nsub, sample2, period, do_auto, do_cross, 
-    estimator, num_threads, max_sample_size):
-    """ 
-    Private method to do bounds-checking on the arguments passed to 
-    `~halotools.mock_observables.jackknife_tpcf`. 
-    """
-    
-    sample1 = convert_to_ndarray(sample1)
-    
-    if sample2 is not None: 
-        sample2 = convert_to_ndarray(sample2)
-        if np.all(sample1==sample2):
-            _sample1_is_sample2 = True
-            msg = ("\n `sample1` and `sample2` are exactly the same, \n"
-                   "only the auto-correlation will be returned.\n")
-            warn(msg)
-            do_cross is False
-        else: 
-            _sample1_is_sample2 = False
-    else: 
-        sample2 = sample1
-        _sample1_is_sample2 = True
-
-    if period is None:
-        PBCs = False
-    else:
-        PBCs = True
-    #process randoms parameter
-    if np.shape(randoms) == (1,):
-        N_randoms = randoms[0]
-        if PBCs is True:
-            randoms = np.random.random((N_randoms,3))*period
-        else:
-            msg = ("\n When no `period` parameter is passed, \n"
-                   "the user must provide true randoms, and \n"
-                   "not just the number of randoms desired.")
-            raise HalotoolsError(msg)
-    
-    # down sample if sample size exceeds max_sample_size.
-    if _sample1_is_sample2 is True:
-        if (len(sample1) > max_sample_size):
-            inds = np.arange(0,len(sample1))
-            np.random.shuffle(inds)
-            inds = inds[0:max_sample_size]
-            sample1 = sample1[inds]
-            msg = ("\n `sample1` exceeds `max_sample_size` \n"
-                   "downsampling `sample1`...")
-            warn(msg)
-    else:
-        if len(sample1) > max_sample_size:
-            inds = np.arange(0,len(sample1))
-            np.random.shuffle(inds)
-            inds = inds[0:max_sample_size]
-            sample1 = sample1[inds]
-            msg = ("\n `sample1` exceeds `max_sample_size` \n"
-                   "downsampling `sample1`...")
-            warn(msg)
-        if len(sample2) > max_sample_size:
-            inds = np.arange(0,len(sample2))
-            np.random.shuffle(inds)
-            inds = inds[0:max_sample_size]
-            sample2 = sample2[inds]
-            msg = ("\n `sample2` exceeds `max_sample_size` \n"
-                   "downsampling `sample2`...")
-            warn(msg)
-    
-    rbins = convert_to_ndarray(rbins)
-    rmax = np.max(rbins)
-    try:
-        assert rbins.ndim == 1
-        assert len(rbins) > 1
-        if len(rbins) > 2:
-            assert array_is_monotonic(rbins, strict = True) == 1
-    except AssertionError:
-        msg = ("\n Input `rbins` must be a monotonically increasing \n"
-               "1-D array with at least two entries")
-        raise HalotoolsError(msg)
-    
-    #Process Nsub entry and check for consistency.
-    Nsub = convert_to_ndarray(Nsub)
-    if len(Nsub) == 1:
-        Nsub = np.array([Nsub[0]]*3)
-    try:
-        assert np.all(Nsub < np.inf)
-        assert np.all(Nsub > 0)
-    except AssertionError:
-        msg = "\n Input `Nsub` must be a bounded positive number in all dimensions"
-        raise HalotoolsError(msg)
-        
-    #Process period entry and check for consistency.
-    if period is None:
-        PBCs = False
-    else:
-        PBCs = True
-        period = convert_to_ndarray(period)
-        if len(period) == 1:
-            period = np.array([period[0]]*3)
-        try:
-            assert np.all(period < np.inf)
-            assert np.all(period > 0)
-        except AssertionError:
-            msg = "\n Input `period` must be a bounded positive number in all dimensions"
-            raise HalotoolsError(msg)
-
-    #check for input parameter consistency
-    if (period is not None):
-        if (rmax >= np.min(period)/3.0):
-            msg = ("\n The maximum length over which you search for pairs of points \n"
-                   "cannot be larger than Lbox/3 in any dimension. \n"
-                   "If you need to count pairs on these length scales, \n"
-                   "you should use a larger simulation. \n")
-            raise HalotoolsError(msg)
-    
-    if (sample2 is not None) & (sample1.shape[-1] != sample2.shape[-1]):
-        msg = ('\n Sample1 and sample2 must have same dimension.')
-        raise HalotoolsError(msg)
-    
-    if (randoms is None) & (PBCs is False):
-        msg = ('\n If no PBCs are specified, randoms must be provided.')
-        raise HalotoolsError(msg)
-    
-    if (type(do_auto) is not bool) | (type(do_cross) is not bool):
-        msg = ('\n `do_auto` and `do_cross` keywords must be of type boolean.')
-        raise HalotoolsError(msg)
-    
-    if num_threads == 'max':
-        num_threads = cpu_count()
-    
-    available_estimators = _list_estimators()
-    if estimator not in available_estimators:
-        msg = ("\n Input `estimator` must be one of the following: \n"
-               "{0}".value(available_estimators))
-        raise HalotoolsError(msg)
-    
-    return sample1, rbins, Nsub, sample2, randoms, period, do_auto, do_cross,\
-           num_threads, _sample1_is_sample2, PBCs
